Create the startupBust class
The setUpGame() method create and initialize the Startup objects with
names and locations. Display brief instructions to the user.

The startPlaying() method that asks the player for guesses and calls the
checkUserGuess() method until all the Startup objects are removed from play.

The checkUserGuess() method that loops through all remaining
Startup objects and calls each Startup object’s checkYourself() method.

The finishGame() method that prints a message about
he user’s performance, based on how many guesses it took
to sink all of the Startup objects

This class still need the GameHelper class in order to compile and\
run successfully.
h